feat: :sparkles: add GetLinksPagination mathod in ormrepo
Also update in sqlrepo method type parametrs int -> uint64

+++ internal/infrastructure/repository/link/sqlrepo/repository.go
@@ -252,7 +252,7 @@ func (r *Repository) GetLinksByTag(ctx context.Context, uid int64, tag string) (
 	return links, nil
 }
 
-func (r *Repository) GetLinksPagination(ctx context.Context, offset, limit int) ([]*domain.Link, error) {
+func (r *Repository) GetLinksPagination(ctx context.Context, offset, limit uint64) ([]*domain.Link, error) {
 	querier := txs.GetQuerier(ctx, r.db)
 
 	query := `
